[Move] 파일 composition 고려하여 폴더 이동 및 리뷰 반영
composition 폴더로 헤더 및 searchbox이동, 리액트스럽게 props 구조분해할당 절대경로 적용 반영
